+++ backend/src/test/java/com/Edu_App/TestData.java
@@ -1,6 +1,7 @@
 package com.Edu_App;
 
 
+import java.math.BigDecimal;
 
 import com.Edu_App.domain.entities.AccountEntity;
 import com.Edu_App.domain.entities.CurrencyEntity;
@@ -30,7 +31,7 @@ public final class TestData {
     {
         return  AccountEntity.builder()
             .iban("BG123456789")
-            .balance(1000.0)
+            .balance(BigDecimal.valueOf(1000.00))
             .owner(userE)
             .currency(currencyE)
             .build();
@@ -40,7 +41,7 @@ public final class TestData {
     {
         return  AccountEntity.builder()
             .iban("BG987654321")
-            .balance(1500.40)
+            .balance(BigDecimal.valueOf(1500.40))
             .owner(userE)
             .currency(currencyE)
             .build();
@@ -50,7 +51,7 @@ public final class TestData {
     {
         return CurrencyEntity.builder()
             .code("BGN")
-            .exchangeRate(1.0)
+            .exchangeRate(BigDecimal.valueOf(1.00))
             .build();
     }
 
@@ -58,14 +59,14 @@ public final class TestData {
     {
         return CurrencyEntity.builder()
             .code("EUR")
-            .exchangeRate(1.97)
+            .exchangeRate(BigDecimal.valueOf(1.97))
             .build();
     }
 
     public static TransferEntity CreateTestTranferEntity1(AccountEntity sender, AccountEntity reciever, CurrencyEntity currencyE)
     {
         return TransferEntity.builder()
-                            .amount(500.0)
+                            .amount(BigDecimal.valueOf(500.0))
                             .sender(sender)
                             .receiver(reciever)
                             .currency(currencyE).build();
@@ -74,7 +75,7 @@ public final class TestData {
     public static TransferEntity CreateTestTranferEntity2(AccountEntity sender, AccountEntity reciever, CurrencyEntity currencyE)
     {
         return TransferEntity.builder()
-                            .amount(250.0)
+                            .amount(BigDecimal.valueOf(250.0))
                             .sender(sender)
                             .receiver(reciever)
                             .currency(currencyE).build();
